chore: Move handler tests to tier tests
